feat(BookedTicketPrecference): 클래스 분리, 패키지정리
