+++ app/src/main/java/woowacourse/movie/model/data/local/preference/TicketIssuePreference.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.model.data.local
+package woowacourse.movie.model.data.local.preference
 
 import android.content.Context
 import android.content.Context.MODE_PRIVATE
