+++ app/src/main/java/woowacourse/movie/presentation/bookedticketlist/BookedTicketsFragment.kt
@@ -7,7 +7,7 @@ import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import androidx.recyclerview.widget.RecyclerView
 import woowacourse.movie.R
-import woowacourse.movie.model.data.local.BookedTicketPreference
+import woowacourse.movie.model.data.local.preference.bookedticket.BookedTicketPreference
 import woowacourse.movie.model.data.remote.DummyMovieStorage
 import woowacourse.movie.presentation.complete.CompleteActivity
 import woowacourse.movie.presentation.model.TicketModel
